+++ make/functions.php
@@ -6,7 +6,7 @@
 /**
  * The current version of the theme.
  */
-define( 'TTFMAKE_VERSION', '1.6.0' );
+define( 'TTFMAKE_VERSION', '1.6.1' );
 
 /**
  * The minimum version of WordPress required for Finder.
@@ -28,6 +28,67 @@ if ( ( defined( 'SCRIPT_DEBUG' ) && true === SCRIPT_DEBUG ) ) {
 if ( ! isset( $content_width ) ) {
 	$content_width = 620;
 }
+
+/**
+ * Load files.
+ *
+ * @since 1.6.1.
+ *
+ * @return void
+ */
+function ttfmake_require_files() {
+	$files = array(
+		// Activation
+		get_template_directory() . '/inc/activation.php',
+		// Compatibility
+		get_template_directory() . '/inc/compatibility.php',
+		// Customizer
+		get_template_directory() . '/inc/customizer/bootstrap.php',
+		// Gallery slider
+		get_template_directory() . '/inc/gallery-slider/gallery-slider.php',
+		// Formatting
+		get_template_directory() . '/inc/formatting/formatting.php',
+		// Integrations
+		get_template_directory() . '/inc/jetpack.php',
+		get_template_directory() . '/inc/woocommerce.php',
+		// Miscellaneous
+		get_template_directory() . '/inc/extras.php',
+		get_template_directory() . '/inc/template-tags.php',
+	);
+
+	if ( is_admin() ) {
+		$admin_files = array(
+			// Admin notices
+			get_template_directory() . '/inc/admin-notice/admin-notice.php',
+			// Page customizations
+			get_template_directory() . '/inc/edit-page.php',
+			// Page Builder
+			get_template_directory() . '/inc/builder/core/base.php'
+		);
+
+		$files = array_merge( $files, $admin_files );
+	}
+
+	/**
+	 * Filter the list of theme files to load.
+	 *
+	 * Note that in some cases, the order that the files are listed in matters.
+	 *
+	 * @since 1.6.1.
+	 *
+	 * @param array    $files    The array of absolute file paths.
+	 */
+	$files = apply_filters( 'make_required_files', $files );
+
+	foreach ( $files as $file ) {
+		if ( file_exists( $file ) ) {
+			require_once $file;
+		}
+	}
+}
+
+// Load files immediately.
+ttfmake_require_files();
 
 if ( ! function_exists( 'ttfmake_content_width' ) ) :
 /**
@@ -71,56 +132,6 @@ function ttfmake_content_width() {
 endif;
 
 add_action( 'template_redirect', 'ttfmake_content_width' );
-
-/**
- * Global includes.
- */
-// Compatibility
-require get_template_directory() . '/inc/compatibility.php';
-
-// Custom functions that act independently of the theme templates
-require get_template_directory() . '/inc/extras.php';
-
-// Custom template tags
-require get_template_directory() . '/inc/template-tags.php';
-
-// Customizer additions
-require get_template_directory() . '/inc/customizer/bootstrap.php';
-
-// Gallery slider
-require get_template_directory() . '/inc/gallery-slider/gallery-slider.php';
-
-// Formatting
-require get_template_directory() . '/inc/formatting/formatting.php';
-
-/**
- * Admin includes.
- */
-if ( is_admin() ) {
-	// Page customizations
-	require get_template_directory() . '/inc/edit-page.php';
-
-	// Page Builder
-	require get_template_directory() . '/inc/builder/core/base.php';
-
-	// Admin notices
-	require get_template_directory() . '/inc/admin-notice/admin-notice.php';
-}
-
-/**
- * 3rd party compatibility includes.
- */
-// Jetpack
-// There are several plugins that duplicate the functionality of various Jetpack modules,
-// so rather than conditionally loading our Jetpack compatibility file based on the presence
-// of the main Jetpack class, we attempt to detect individual classes/functions related to
-// their particular modules.
-require get_template_directory() . '/inc/jetpack.php';
-
-// WooCommerce
-if ( class_exists( 'WooCommerce' ) ) {
-	require get_template_directory() . '/inc/woocommerce.php';
-}
 
 if ( ! function_exists( 'ttfmake_setup' ) ) :
 /**
@@ -374,12 +385,12 @@ function ttfmake_scripts() {
 		true
 	);
 
+	// Cycle2
+	// Register only. Enqueued when necessary by gallery shortcode or Banner section.
+	ttfmake_cycle2_script_setup( array( 'jquery' ) );
+
 	// jQuery
 	$script_dependencies[] = 'jquery';
-
-	// Cycle2
-	ttfmake_cycle2_script_setup( $script_dependencies );
-	$script_dependencies[] = 'ttfmake-cycle2';
 
 	// Global script
 	wp_enqueue_script(
@@ -447,7 +458,7 @@ if ( ! function_exists( 'ttfmake_cycle2_script_setup' ) ) :
  */
 function ttfmake_cycle2_script_setup( $script_dependencies ) {
 	if ( defined( 'TTFMAKE_SUFFIX' ) && '.min' === TTFMAKE_SUFFIX ) {
-		wp_enqueue_script(
+		wp_register_script(
 			'ttfmake-cycle2',
 			get_template_directory_uri() . '/js/libs/cycle2/jquery.cycle2' . TTFMAKE_SUFFIX . '.js',
 			$script_dependencies,
@@ -456,7 +467,7 @@ function ttfmake_cycle2_script_setup( $script_dependencies ) {
 		);
 	} else {
 		// Core script
-		wp_enqueue_script(
+		wp_register_script(
 			'ttfmake-cycle2',
 			get_template_directory_uri() . '/js/libs/cycle2/jquery.cycle2.js',
 			$script_dependencies,
@@ -465,7 +476,7 @@ function ttfmake_cycle2_script_setup( $script_dependencies ) {
 		);
 
 		// Vertical centering
-		wp_enqueue_script(
+		wp_register_script(
 			'ttfmake-cycle2-center',
 			get_template_directory_uri() . '/js/libs/cycle2/jquery.cycle2.center.js',
 			'ttfmake-cycle2',
@@ -474,7 +485,7 @@ function ttfmake_cycle2_script_setup( $script_dependencies ) {
 		);
 
 		// Swipe support
-		wp_enqueue_script(
+		wp_register_script(
 			'ttfmake-cycle2-swipe',
 			get_template_directory_uri() . '/js/libs/cycle2/jquery.cycle2.swipe.js',
 			'ttfmake-cycle2',